refactor(web): extract useAbortableMount hook
Consolidate the mountedRef + abortRef + cleanup pattern shared by
ChatForm and RunForm into lib/hooks/useAbortableMount.ts.
